Incorporated slides and website stylesheets into the build.
Note: This currently only affects the "distrib" (doc) build. So if
you don't need to build distrib/doc, you won't be affected by this
change. If you DO need to build distrib/doc, it will break unless
you use the xsl/Makefile from the "build" branch instead of from
the head.

This build alters the distrib build such that:

  - an xsl/slides directory is created by copying over the
    contents of the slides/xsl

  - an xsl/website directory is created by copying over the
    contents of the website/xsl directory

  - the reference.html part of the doc build now adds the slides
    and website param reference doc

This is an experiment. If we decide to go ahead with it in the
release build, and everything is found to be OK when it gets out
to users and they test it, then the next step would be to ask SF
admin to move the website/xsl and slides/xsl CVS directory into
xsl/ to create xsl/slides and xsl/website, and they would be
maintained in the xsl/ CVS going forward.


git-svn-id: https://docbook.svn.sourceforge.net/svnroot/docbook/branches/unlabeled-1.21.2@4924 6cfa91a5-b114-0410-aa58-ce7c4e7541e0