Added initial, EXPERIMENTAL support for generating content for HTML "title" attributes from content of the Alt element.
This change adds support for generating HTML "title" attributes
for the following inline elements only (all inlines -- support for
block elements will need to wait for later).

  abbrev accel acronym action application authorinitials beginpage
  citation citerefentry citetitle city classname code command
  comment computeroutput constant country database email envar
  errorcode errorname errortext errortype exceptionname fax
  filename firstname firstterm foreignphrase function glossterm
  guibutton guiicon guilabel guimenu guimenuitem guisubmenu
  hardware honorific interface interfacedefinition interfacename
  keycap keycode keysym lineage lineannotation literal markup
  medialabel methodname mousebutton option optional otheraddr
  othername package parameter personname phone pob postcode
  productname productnumber prompt property quote refentrytitle
  remark replaceable returnvalue sgmltag shortcut state street
  structfield structname subscript superscript surname symbol
  systemitem tag termdef token trademark type uri userinput
  varname wordasword

Implemented by creating a new named template, generate.html.title.
That is is currently called in eleven places only, in the
inline.xsl file. But it's called by all the inline.* templates
(e.g., inline.boldseq), which in turn are called by other
(element) templates, so it results, currently, in supporting
generation of the HTML "title" attribute for a total of about 93
elements (the list above).


git-svn-id: https://docbook.svn.sourceforge.net/svnroot/docbook/trunk@4942 6cfa91a5-b114-0410-aa58-ce7c4e7541e0