Replace hyphens with "\-".
<sigh/> Although the old-school gnomes who maintain groff
apparently don't figure it's necessary to document this anywhere,
there is apparently only one reliable way to make a non-breaking
hyphen in groff: with "\-". At least that it what testing in my
environment seems to indicate.

And because groff is not at all smart in the way it breaks things
at hyphens (it will, for example, break command arguments across
lines if they contain hyphens), we must globally replace all
hyphens with "\-".

God in heaven, please let me live to see a day when roff and all
is (r)offspring have disappeared from the face of the earth.


git-svn-id: https://docbook.svn.sourceforge.net/svnroot/docbook/trunk/xsl@4976 6cfa91a5-b114-0410-aa58-ce7c4e7541e0