Ulink handling -- fixed performance issue (per-Refentry sets of Ulinks are now assembled once per document and stored in a global variable).
It now takes roughly on the order of 15 to 20 percent longer to
process a doc with links numbered than it does to process it with
links unnumbered. Which may sound like a lot, but with a doc that
has 100+ links, on my machine, that still amounts to less than a
one-second difference (only about 0.7 seconds).


git-svn-id: https://docbook.svn.sourceforge.net/svnroot/docbook/trunk@5080 6cfa91a5-b114-0410-aa58-ce7c4e7541e0