Implemented out-of-line handling of URLs for links (Ulinks). (Closes #1234711 and #1233909)
For display of links, this implementation gives users three choices:

  1. Number & list links. Each link is numbered and a numbered
     list of all links is added to the end of the document.

     or

  2. Only list links. Links are not numbered, but an (unnumbered)
     list of links is added to the end of the document.

     or

  3. Suppress links. Don't number links and don't add any list of
     links to the end of the document

Users can also choose whether links should be underlined.

Default is "the works" -- list, number, and underline links. The
man.links.* parameters can be used to change the defaults.

The default heading for the link list is REFERENCES. That can be
changed using the man.links.list.heading param.

There is a performance hit for listing and numbering links. But it
is not an unreasonable one.

The code currently only checks for Ulinks. It can be updated later
(possibly) to deal with DocBook NG linking. But come to think of
it, DocBook NG links get converted by the strip-ns template. So
DocBook NG users will just need to run that first. Or, if I can
ever manage to get the single-pass conversion process to work with
manpages, it will get done automatically.


git-svn-id: https://docbook.svn.sourceforge.net/svnroot/docbook/trunk/xsl@5086 6cfa91a5-b114-0410-aa58-ce7c4e7541e0