Fixed bug that caused literal .sp to show up in output. (Thanks to Costin Stroie for reporting).
WARNING: One of the consequences of this change is that the
stylesheet no longer generates spaces at the end of certain lines.
I had it doing that as a sort of hack to prevent certain kinds of
problems in output. It seems like the hack my no longer be
necessary. But if I am wrong, this change may cause spaces to show
up at the beginning of some lines in "mixed content". If you see
those, then this change has introduced a regression, and I will
need to go back and figure out the right way to fix it (instead of
the kludgy way I was doing it before.

Also, instances like the following were showing up in some cases.

  .PP
  .sp

That is, a .PP macro immediately followed by a .sp macro. I can
think of no good read to ever generate that, and I am not sure why
it is getting generated. So I have taked the lazy way out and
caused the stylesheet to now strip out the .sp from all such
instances, leaving just the .PP


git-svn-id: https://docbook.svn.sourceforge.net/svnroot/docbook/trunk/xsl@5140 6cfa91a5-b114-0410-aa58-ce7c4e7541e0