Added a <div class="{$class}-contents"> wrapper around output of contents of all formal objects. Also, added an optional <br class="{class}-break"/> linebreak after all formal objects.
WARNING: Because this change places an additional DIV between the
DIV wrapper for the equation and the equation contents, it may
break some existing CSS stylesheets that have been created with
the assumption that there would never be an intervening DIV there.

The following is an example of what Equation output looks like as
a result of the changes described above.

  <div class="equation">
    <a name="three" id="three"></a>

    <p class="title"><b>(1.3)</b></p>

    <div class="equation-contents">
      <span class="mathphrase">1+1=3</span>
    </div>
  </div><br class="equation-break">

Rationale: These changes allow CSS control of the placement of the
formal-object title relative to the formal-object contents. For
example, using the CSS "float" property enables the title and
contents to be rendered on the same line. Example stylesheet:

  .equation {
    margin-top: 20px;
    margin-bottom: 20px;
  }
  .equation-contents {
    float: left;
  }

  .equation .title {
    margin-top: 0;
    float: right;
    margin-right: 200px;
  }

  .equation .title b {
    font-weight: normal;
  }

  .equation-break {
    clear: both;
  }

Note that the purpose of the ".equation-break" class is to provide
a way to clear off the floats.

If you want to instead have the equation title rendered to the left
of the equation contents, you can do something like this:

  .equation {
    margin-top: 20px;
    width: 300px;
    margin-bottom: 20px;
  }
  .equation-contents {
    float: right;
  }

  .equation .title {
    margin-top: 0;
    float: left;
    margin-right: 200px;
  }

  .equation .title b {
    font-weight: normal;
  }

  .equation-break {
    clear: both;
  }


git-svn-id: https://docbook.svn.sourceforge.net/svnroot/docbook/trunk@5471 6cfa91a5-b114-0410-aa58-ce7c4e7541e0