Added the man.charmap.subset.profile.english parameter and refined the handling of charmap subsets to differentiate between English and non-English source.
This way charmap subsets are now handled is this:

If the value of the man.charmap.use.subset parameter is non-zero,
and your DocBook source is not written in English (that is, if its
lang or xml:lang attribute has a value other than en), then the
character-map subset specified by the man.charmap.subset.profile
parameter is used instead of the full roff character map.

Otherwise, if the lang or xml:lang attribute on the root element
in your DocBook source or on the first refentry element in your
source has the value en or if it has no lang or xml:lang
attribute, then the character-map subset specified by the
man.charmap.subset.profile.english parameter is used instead of
man.charmap.subset.profile.

The difference between the two subsets is that
man.charmap.subset.profile provides mappings for characters in
Western European languages that are not part of the Roman
(English) alphabet (ASCII character set).


git-svn-id: https://docbook.svn.sourceforge.net/svnroot/docbook/trunk@7743 6cfa91a5-b114-0410-aa58-ce7c4e7541e0